changed data type

+++ src/main/java/numberTranslation.java
@@ -1,6 +1,7 @@
 import service.NumberTranslationService;
 import service.impl.NumberTranslationServiceImpl;
 ;
+import java.math.BigInteger;
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
@@ -14,10 +15,14 @@ public class numberTranslation {
         while (true) {
 
             try {
+
                 Scanner in = new Scanner(System.in);
+
                 System.out.print("Введите число: ");
-                double num = in.nextDouble();
+                BigInteger num = in.nextBigInteger();
                 String number = numberTranslationService.NumberTranslate(num);
+
+
                 System.out.println("Число в текстовом виде: " + number);
 
             } catch (InputMismatchException e) {
